% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceFeatures.R
\name{sequenceFeatures}
\alias{sequenceFeatures}
\title{Generate the sequence feature matrix from GRanges.}
\usage{
sequenceFeatures(query_gr, bsgnm)
}
\arguments{
\item{query_gr}{a \code{GRanges} object to generate the underlying sequence features, the input ranges must be equal in sizes.}

\item{bsgnm}{a \code{BSgenome} object which encode the genome information.}
}
\value{
a data.frame contains the sequence features in its collumns.
}
\description{
\code{sequenceFeatures} is used to extract the sequence features given GenomicRanges object.
The sequence feature is composed by the physical chemical properties and the accumulative frequencies of DNA nucleotides [1].
}
\examples{
## --------------------------------------------------------------------
## Sequence features for hg19
## --------------------------------------------------------------------
\dontrun{
library(BSgenome.Hsapiens.UCSC.hg19)

feature_hg19 <- sequenceFeatures(
query_gr = m6A_ex_hg19,
bsgnm = Hsapiens
)
}

}
\references{
1. Bari, A.T.M.G., et al., DNA Encoding for Splice Site Prediction in Large DNA Sequence. 2013: Springer Berlin Heidelberg. 46-58.
}
