% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptomicFeatures.R
\name{transcriptomicFeatures}
\alias{transcriptomicFeatures}
\title{Generation of transcriptomic features for range based epitranscriptomic dataset}
\usage{
transcriptomicFeatures(
  mod_gr,
  txdb,
  bsgnm = NULL,
  fc = NULL,
  pc = NULL,
  struct_hybridize = NULL,
  genomic_annot = NULL,
  motif = NULL,
  annot_clustering = NULL,
  motif_clustering = NULL,
  hk_genes_id = NULL,
  miRtar_genes_id = NULL,
  isoform_ambiguity_method = c("longest_tx", "average"),
  gene_ambiguity_method = c("average", "drop_overlap"),
  standardization = FALSE
)
}
\arguments{
\item{mod_gr}{A \code{GRanges} object for the genomic location of RNA modification sites.}

\item{txdb}{A \code{TxDb} object for the transcript annotation.

The \code{TxDb} can be obtained from either Bioconductor or from the GFF/GTF files using the function \code{\link{makeTxDbFromGFF}}.}

\item{bsgnm}{Optional; a \code{BSgenome} object for the genomic sequence annotation.

The \code{BSgenome} can be downloaded either from Bioconductor or through the function \code{\link{getBSgenome}}.}

\item{fc, pc}{Optional; \code{GScores} objects for the annotation of phastCons[1] scores and the fitness consequences scores[2].

The \code{GSscores} can be downloaded from Bioconductor.}

\item{struct_hybridize}{A \code{GRanges} object recording the hybridized region of the RNA transcripts.

WHISTLE includes the predicted exonic RNA secondary structures of hg19 and mm10 in the exported data.
Please check \code{\link{Struc_hg19}} and \code{\link{Struc_mm10}} for details.}

\item{genomic_annot}{A \code{list} of \code{\link{GRanges}} for the user defined range based functional genomic annotation.

New dummy variable features will be created for each list element by overlapping the \code{GRanges} with the corresponding range based annotations.
The names of the list will correspond to the names of features in the returned \code{data.frame} object.

Some of the range based functional genomic annotations can be important in the site prediction of RNA modifications.
These annotations usually include the binding sites of the regulator proteins (e.x. writers, erasers, readers) related to the RNA modification.
Additionally, the microRNA binding targets can associate with the locations of certain RNA modification.

WHISTLE package includes the following build-in \code{list} objects to represent the RBP binding sites of the regulator proteins for different RNA modifications:
\code{\link{RBP_m6A_hg19}}, \code{\link{RBP_m6A_mm10}}, \code{\link{RBP_m1A_hg19}}, \code{\link{RBP_m5C_hg19}}, \code{\link{RBP_Psi_hg19}}; all of the RBP data in the list objects are gathered from Starbase[5].

Furthermore, the information of microRNA binding sites could be found in :\code{\link{TargetScan_hg19}}.}

\item{motif}{A character vector for the sequence motifs centered by the RNA modification site.
Each element of the character vector will become an annotated feature in the returned table.

For example, to predict m6A RNA modification, the motif vector could be the combination of the instances in RRACH.

P.S. The motif will not be attached if the \code{mod_gr} is not in single nucleotide resolution (with all width = 1).}

\item{annot_clustering}{A \code{GRanges} object to generate clustering indexes of the modification sites related to a particular range based annotation;
under the default setting, the clustering will be conducted on the input GRanges.

4 features --- \code{clust_f100}, \code{clust_f1000}, \code{dist_nearest_p200}, and \code{dist_nearest_p2000} will be generated to represent the clustering effect of the provided annotation related to the RNA modification site.}

\item{motif_clustering}{A \code{character} indicating the motif used when generating the features for the motif clustering; default: "DRACH".

The motif clustering will represent the clustering of the sequence motif using 4 features similar to the annotation clusering.}

\item{hk_genes_id}{Optional; A character string for the gene id of the house keeping genes. The ids should correspond to the gene ids in the provided \code{TxDb} object.

The entrez gene IDs of the house keeping genes for hg19 are included in WHISTLE: \code{\link{HK_hg19_eids}} [6].}

\item{miRtar_genes_id}{Optional; A character string for the Gene IDs of the microRNA targeted genes. The id should also correspond to the ids in the provided \code{TxDb} object.

The entrez gene IDs of the microRNA targeted genes for hg19 are included in WHISTLE: \code{\link{miRtarGene_eid_hg19}}.}

\item{isoform_ambiguity_method}{Can be \code{"longest_tx"} or \code{"average"}. The former keeps only the longest transcript as the transcript annotation.
The later will use the average feature entries for multiple mappings of the transcript isoform; default: \code{"longest_tx"}.}

\item{gene_ambiguity_method}{Can be \code{"drop_overlap"} or \code{"average"}. The former will not annotate the modification Site overlapped with > 1 genes (By returning NA).
The later will use the average feature entries for mapping of multiple genes; default: \code{"average"}.}

\item{standardization}{A logical indicating whether to standardize the continous features. For a given scalar feature X, the standardization is conducted by \eqn{ \frac{X - mean(X)}{sd(X)} }{(X - mean(X))/sd(X)}; default: \code{FALSE}.}
}
\value{
The function accepts the input of a \code{GRanges} object for RNA modification sites, and it will return a \code{data.frame} with columns of the annotated transcriptomic features.
The rows of the returned \code{data.frame} object will correspond to the order of sites in the corresponding GRanges.
}
\description{
The function \bold{\code{transcriptomicFeatures}} could generate features based on the \code{\link{GRanges}} input object of the sites for RNA epitranscriptomic modification.
Users could provide additional genomic and transcripomic annotations to expand the feature sets annotated by this function.

Under the default setting, \code{transcriptomicFeatures} could provide 40 topological features based on the \code{\link{TxDb}} object of the transcript annotation.
More than 80 features related to the domain knowledge of epitranscriptomic modifications could be annotated given a complete support of the annotation objects. The suggested additonal annotation data types include:

\enumerate{
\item \code{\link{BSgenome}} object for the genome sequence information, such as the \code{BSgenome.Hsapiens.UCSC.hg19} package that provides the genomic sequences for the hg19 genome assembly.

\item \code{\link{GScores}} objects for the annotation of evolutionary conservation scores, such as the \code{phastCons100way.UCSC.hg19} package that provides the annotation of Phastcons scores for hg19.

\item \code{GRanges} or \code{\link{GRangesList}} objects that providing the annotations of other range based transcriptomic features gathered from diffrent databases.
}

Please download the suggested annotation packages from \href{https://www.bioconductor.org/}{Bioconductor}.
}
\details{
\code{transcriptomicFeatures} in \bold{WHISTLE} package can retrieve a set of domain knowledge based transcriptomic features that contribute toward the topology of epitranscriptomic modifications.

The utility of the annotated transcriptomic features is 2 folds:

\enumerate{
\item The transcriptomic features are useful in the machine learning prediction of the epitranscriptomic markers.
WHISTLE predictor utilizes part of the transcriptomic features to achieve superior performances on the m6A site preidiction than any other predictors based only on the sequence features[3].

\item The features are also important for the topological characterization of the epitranscriptomic regulations under a new biological condition.
The statistical association (e.x. with linear or logistic regression) between the transcriptomic feature and the RNA modification measurements can elucidate the scientific significance of a particular feature in epitranscriptomic regulations[4].

}

In general, \code{transcriptomicFeatures} could annotate the transcriptomic features that belong to the following categories:

\describe{
\item{\strong{Topological features}}{Features constructed from the transcriptomic landmarks. The toplogical features were previously reported by the published studies to associate with the locations of RNA modifiations.}

\item{\strong{Sequence features}}{The sequence derived features that are related to epitranscriptomics, such as the sequence motifs, predicted RNA secondary structures, and GC contents.}

\item{\strong{Functional genomic features}}{Features of other functional genomic information that have proofed correlation with the mRNA modifications,
such as the functional aspects of the transcript, evolutionary conservation of the sequence, and the RBP binding sites of RNA modification writers, readers, and erasers. }
}

By default, only the topological features will be annotated if users provide the \code{TxDb} object along without additional annotation objects.
}
\examples{

## --------------------------------------------------------------------
## Topological features for hg19
## --------------------------------------------------------------------

library(TxDb.Hsapiens.UCSC.hg19.knownGene)

txdb_hg19 <- TxDb.Hsapiens.UCSC.hg19.knownGene

feature_hg19 <- transcriptomicFeatures(
mod_gr = m6A_ex_hg19,
txdb = txdb_hg19
)

str(feature_hg19)

## --------------------------------------------------------------------
## Topological features for mm10
## --------------------------------------------------------------------
\dontrun{
library(TxDb.Mmusculus.UCSC.mm10.knownGene)

txdb_mm10 <- TxDb.Mmusculus.UCSC.mm10.knownGene

feature_mm10 <- transcriptomicFeatures(
mod_gr = m6A_ex_mm10,
txdb = txdb_mm10
)

str(feature_mm10)

## --------------------------------------------------------------------
## Complete features for hg19
## --------------------------------------------------------------------

library(TxDb.Hsapiens.UCSC.hg19.knownGene)
library(BSgenome.Hsapiens.UCSC.hg19)
library(fitCons.UCSC.hg19)
library(phastCons100way.UCSC.hg19)

txdb_hg19 <- TxDb.Hsapiens.UCSC.hg19.knownGene

feature_complete <- transcriptomicFeatures(

   mod_gr = m6A_ex_hg19,
# TxDb object for the transcript annotation.
   txdb = txdb_hg19,

# BSgenome object for the genome sequence.
   bsgnm = Hsapiens,

# GScores object for the Fitness Consequence scores.
   fc = fitCons.UCSC.hg19,

# GScores object for the PhastCons scores.
   pc = phastCons100way.UCSC.hg19,

# RNA fold predicted RNA secondary structures for hg19 exons.
   struct_hybridize = Struc_hg19,

# Binding sites of m6A regulators.
   genomic_annot = RBP_m6A_hg19,

# Entrez gene ids of house keeping genes.
   hk_genes_id = hkGene_eid_hg19,

# Entrez gene ids of microRNA targeted genes.
   miRtar_genes_id = miRtarGene_eid_hg19,

# Instances of DRACH motif for m6A modification.
   motif = c("AAACA","AGACA","AAACT","AGACT","AAACC","AGACC",
             "GAACA","GGACA","GAACT","GGACT","GAACC","GGACC",
             "TAACA","TGACA","TAACT","TGACT","TAACC","TGACC"),

# The clustering effect on DRACH motifs.
   motif_clustering = "DRACH",

# The clustering on the input modification Granges.
   annot_clustering = m6A_ex_hg19
)

str(feature_complete)

}
}
\references{
\enumerate{

\item Siepel A, Bejerano G, Pedersen JS, Hinrichs AS, Hou M, Rosenbloom K, Clawson H, Spieth J, Hillier LW, Richards S: Evolutionarily conserved elements in vertebrate, insect, worm, and yeast genomes. Genome research 2005, 15(8):1034-1050.

\item Gulko B, Hubisz MJ, Gronau I, Siepel A: A method for calculating probabilities of fitness consequences for point mutations across the human genome. Nature genetics 2015, 47(3):276.

\item Chen K, Wei Z, Zhang Q, Wu X, Rong R, Lu Z, Su J, de Magalhaes JP, Rigden DJ, Meng J: WHISTLE: a high-accuracy map of the human N6-methyladenosine (m6A) epitranscriptome predicted using a machine learning approach. Nucleic Acids Research 2019.

\item Schwartz S, Mumbach M, Jovanovic M, Wang T, Maciag K, Bushkin GG, Mertins P, Ter-Ovanesyan D, Habib N, Cacchiarelli D: Perturbation of m6A Writers Reveals Two Distinct Classes of mRNA Methylation at Internal and 5' Sites. Cell Reports 2014, 8(1):284-296.

\item Li J-H, Liu S, Zhou H, Qu L-H, Yang J-H: starBase v2.0: decoding miRNA-ceRNA, miRNA-ncRNA and protein–RNA interaction networks from large-scale CLIP-Seq data. Nucleic acids research 2013, 42(D1):D92-D97.

\item Eisenberg E, Levanon EY: Human housekeeping genes, revisited. Trends in Genetics 2013, 29(10):569-574.

}
}
